requirements.txt
feedparser
pyyaml
google-genai
Pillow
